Use correct size type for C FFI

+++ rust-bits/src/lib.rs
@@ -29,9 +29,9 @@ pub extern "C" fn signature_public_key_compress(pk: &signature::PublicKey, out: 
 #[no_mangle]
 pub extern "C" fn signature_public_key_decompress(
     data: *const u8,
-    data_len: i32,
+    data_len: usize,
 ) -> Option<Box<signature::PublicKey>> {
-    let data = unsafe { slice::from_raw_parts(data, data_len as usize) };
+    let data = unsafe { slice::from_raw_parts(data, data_len) };
     match signature::PublicKey::try_from(data) {
         Err(_) => None,
         Ok(pk) => Some(Box::new(pk)),
@@ -42,11 +42,11 @@ pub extern "C" fn signature_public_key_decompress(
 pub extern "C" fn signature_sign(
     priv_ptr: *const u8,
     data: *const u8,
-    data_len: i32,
+    data_len: usize,
     out: *mut u8,
 ) {
     let private = unsafe { signature::PrivateKey::from_pointer(priv_ptr) };
-    let message = unsafe { slice::from_raw_parts(data, data_len as usize) };
+    let message = unsafe { slice::from_raw_parts(data, data_len) };
     let sig_bytes: [u8; signature::SIGNATURE_SIZE] = private.sign(message).into();
     unsafe { out.copy_from(sig_bytes.as_ptr(), sig_bytes.len()) }
 }
@@ -55,10 +55,10 @@ pub extern "C" fn signature_sign(
 pub extern "C" fn signature_verify(
     public: &signature::PublicKey,
     data: *const u8,
-    data_len: i32,
+    data_len: usize,
     sig_ptr: *const u8,
 ) -> bool {
-    let message = unsafe { slice::from_raw_parts(data, data_len as usize) };
+    let message = unsafe { slice::from_raw_parts(data, data_len) };
     let sig = unsafe { signature::Signature::from_pointer(sig_ptr) };
     public.verify(&sig, message)
 }
